begin refactoring templates
